requirement.txt
pip install nltk
nltk.download('punkt')
nltk.download('stopwords')

